Drop ResourceProxy; #290
